Refactor PlayerRepository into several smaller repositories
