+++ SmbExplorerCompanion.Core/Commands/Queries/Players/GetHallOfFameCandidatesRequest.cs
@@ -19,15 +19,46 @@ public class GetHallOfFameCandidatesRequest : IRequest<OneOf<RetiredPlayerCareer
     internal class GetHallOfFameCandidatesHandler : IRequestHandler<GetHallOfFameCandidatesRequest,
         OneOf<RetiredPlayerCareerStatsDto, None, Exception>>
     {
-        private readonly IPlayerRepository _playerRepository;
+        private readonly IPositionPlayerCareerRepository _positionPlayerCareerRepository;
+        private readonly IPitcherCareerRepository _pitcherCareerRepository;
 
-        public GetHallOfFameCandidatesHandler(IPlayerRepository playerRepository)
+        public GetHallOfFameCandidatesHandler(IPositionPlayerCareerRepository positionPlayerCareerRepository,
+            IPitcherCareerRepository pitcherCareerRepository)
         {
-            _playerRepository = playerRepository;
+            _positionPlayerCareerRepository = positionPlayerCareerRepository;
+            _pitcherCareerRepository = pitcherCareerRepository;
         }
 
         public async Task<OneOf<RetiredPlayerCareerStatsDto, None, Exception>> Handle(GetHallOfFameCandidatesRequest request,
-            CancellationToken cancellationToken) =>
-            await _playerRepository.GetHallOfFameCandidates(request.SeasonId, cancellationToken);
+            CancellationToken cancellationToken)
+        {
+            var positionPlayersResponse = await _positionPlayerCareerRepository.GetHallOfFameCandidates(request.SeasonId, cancellationToken);
+            if (positionPlayersResponse.TryPickT2(out var exception, out var rest))
+            {
+                return exception;
+            }
+
+            if (rest.TryPickT1(out var none, out var positionPlayers))
+            {
+                return none;
+            }
+
+            var pitchersResponse = await _pitcherCareerRepository.GetHallOfFameCandidates(request.SeasonId, cancellationToken);
+            if (pitchersResponse.TryPickT2(out exception, out var rest2))
+            {
+                return exception;
+            }
+
+            if (rest2.TryPickT1(out none, out var pitchers))
+            {
+                return none;
+            }
+
+            return new RetiredPlayerCareerStatsDto
+            {
+                BattingCareers = positionPlayers,
+                PitchingCareers = pitchers
+            };
+        }
     }
 }
