+++ SmbExplorerCompanion.Core/Interfaces/IGeneralPlayerRepository.cs
@@ -1,0 +1,20 @@
+﻿using SmbExplorerCompanion.Core.Entities.Players;
+using OneOf;
+
+namespace SmbExplorerCompanion.Core.Interfaces;
+
+public interface IGeneralPlayerRepository
+{
+    public Task<OneOf<PlayerOverviewDto, Exception>> GetHistoricalPlayer(int playerId,
+        CancellationToken cancellationToken = default);
+
+    public Task<OneOf<List<PlayerFieldingRankingDto>, Exception>> GetPlayerFieldingRankings(int seasonId,
+        int? primaryPositionId,
+        int? pageNumber,
+        int? limit,
+        CancellationToken cancellationToken = default);
+
+    public Task<OneOf<GameStatDto, Exception>> GetLeagueAverageGameStats(int seasonId, bool isPitcher, CancellationToken cancellationToken = default);
+    
+    public Task<OneOf<PlayerGameStatPercentileDto, Exception>> GetPlayerGameStatPercentiles(int playerId, int seasonId, bool isPitcher, CancellationToken cancellationToken = default);
+}
