+++ SmbExplorerCompanion.Database/DependencyInjection.cs
@@ -6,6 +6,9 @@ using SmbExplorerCompanion.Core.Entities.Seasons;
 using SmbExplorerCompanion.Core.Interfaces;
 using SmbExplorerCompanion.Csv.Services;
 using SmbExplorerCompanion.Database.Services;
+using SmbExplorerCompanion.Database.Services.Imports;
+using SmbExplorerCompanion.Database.Services.Lookups;
+using SmbExplorerCompanion.Database.Services.Players;
 
 namespace SmbExplorerCompanion.Database;
 
@@ -24,7 +27,11 @@ public static class DependencyInjection
             .AddScoped<IRepository<PlayerAwardDto>, PlayerAwardRepository>()
             .AddScoped<IRepository<PitcherRoleDto>, PitcherRoleRepository>()
             .AddScoped<ITeamRepository, TeamRepository>()
-            .AddScoped<IPlayerRepository, PlayerRepository>()
+            .AddScoped<IGeneralPlayerRepository, GeneralPlayerRepository>()
+            .AddScoped<IPitcherCareerRepository, PitcherCareerRepository>()
+            .AddScoped<IPitcherSeasonRepository, PitcherSeasonRepository>()
+            .AddScoped<IPositionPlayerCareerRepository, PositionPlayerCareerRepository>()
+            .AddScoped<IPositionPlayerSeasonRepository, PositionPlayerSeasonRepository>()
             .AddScoped<IAwardDelegationRepository, AwardDelegationRepository>()
             .AddScoped<ISearchRepository, SearchRepository>()
             .AddScoped<ISummaryRepository, SummaryRepository>()
