+++ SmbExplorerCompanion.Database/Services/Lookups/FranchiseRepository.cs
@@ -1,12 +1,11 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using OneOf;
 using OneOf.Types;
-using SmbExplorerCompanion.Core.Entities;
 using SmbExplorerCompanion.Core.Entities.Franchises;
 using SmbExplorerCompanion.Core.Interfaces;
 using SmbExplorerCompanion.Database.Mappings;
 
-namespace SmbExplorerCompanion.Database.Services;
+namespace SmbExplorerCompanion.Database.Services.Lookups;
 
 public class FranchiseRepository : IRepository<FranchiseDto>
 {
